src/requirements.txt
numpy==1.17.4
torch==1.3.1
matplotlib==3.1.2
torchvision==0.4.2
pyyaml==5.1.2
tensorboard-pytorch==0.7.1
